Fix: relative paths in @import statements don't work.
Previous attempts to allow relative paths in @import statements added
to `sassOptions.includePaths`. That was deemed unacceptable as it would
capture relative files that weren't explicitly imported with relative paths.

The solution is to pass the filename in `sassOptions`. It's fine to pass
the javascript file path: we are already providing the `data` option with
the content to parse, so sass won't actually try to read the file. It is only
used for resolving explicit relative imports.